+++ plugins/nf-schema/src/main/nextflow/validation/config/HelpConfig.groovy
@@ -2,8 +2,7 @@ package nextflow.validation.config
 
 import groovy.util.logging.Slf4j
 
-import nextflow.validation.Utils
-
+import static nextflow.validation.utils.Colors.removeColors
 /**
  * This class allows to model a specific configuration, extracting values from a map and converting
  *
@@ -32,9 +31,9 @@ class HelpConfig {
         fullParameter       = config.fullParameter              ?: "helpFull"
         showHiddenParameter = config.showHiddenParameter        ?: "showHidden"
         if (monochromeLogs) {
-            beforeText  = config.beforeText ? Utils.removeColors(config.beforeText): ""
-            afterText   = config.afterText  ? Utils.removeColors(config.afterText) : ""
-            command     = config.command    ? Utils.removeColors(config.command)   : ""
+            beforeText  = config.beforeText ? removeColors(config.beforeText): ""
+            afterText   = config.afterText  ? removeColors(config.afterText) : ""
+            command     = config.command    ? removeColors(config.command)   : ""
         } else {
             beforeText  = config.beforeText ?: ""
             afterText   = config.afterText  ?: ""
