+++ plugins/nf-schema/src/main/nextflow/validation/utils/Colors.groovy
@@ -1,0 +1,96 @@
+package nextflow.validation.utils
+
+import groovy.util.logging.Slf4j
+
+/**
+ * A collection of functions for everything color related
+ *
+ * @author : mirpedrol <mirp.julia@gmail.com>
+ * @author : nvnieuwk <nicolas.vannieuwkerke@ugent.be>
+ * @author : KevinMenden
+ */
+
+@Slf4j
+public class Colors {
+
+    //
+    // ANSII Colours used for terminal logging
+    //
+    public static Map logColors(Boolean monochrome_logs) {
+        Map colorcodes = [:]
+
+        // Reset / Meta
+        colorcodes['reset']      = monochrome_logs ? '' : "\033[0m"
+        colorcodes['bold']       = monochrome_logs ? '' : "\033[1m"
+        colorcodes['dim']        = monochrome_logs ? '' : "\033[2m"
+        colorcodes['underlined'] = monochrome_logs ? '' : "\033[4m"
+        colorcodes['blink']      = monochrome_logs ? '' : "\033[5m"
+        colorcodes['reverse']    = monochrome_logs ? '' : "\033[7m"
+        colorcodes['hidden']     = monochrome_logs ? '' : "\033[8m"
+
+        // Regular Colors
+        colorcodes['black']      = monochrome_logs ? '' : "\033[0;30m"
+        colorcodes['red']        = monochrome_logs ? '' : "\033[0;31m"
+        colorcodes['green']      = monochrome_logs ? '' : "\033[0;32m"
+        colorcodes['yellow']     = monochrome_logs ? '' : "\033[0;33m"
+        colorcodes['blue']       = monochrome_logs ? '' : "\033[0;34m"
+        colorcodes['purple']     = monochrome_logs ? '' : "\033[0;35m"
+        colorcodes['cyan']       = monochrome_logs ? '' : "\033[0;36m"
+        colorcodes['white']      = monochrome_logs ? '' : "\033[0;37m"
+
+        // Bold
+        colorcodes['bblack']     = monochrome_logs ? '' : "\033[1;30m"
+        colorcodes['bred']       = monochrome_logs ? '' : "\033[1;31m"
+        colorcodes['bgreen']     = monochrome_logs ? '' : "\033[1;32m"
+        colorcodes['byellow']    = monochrome_logs ? '' : "\033[1;33m"
+        colorcodes['bblue']      = monochrome_logs ? '' : "\033[1;34m"
+        colorcodes['bpurple']    = monochrome_logs ? '' : "\033[1;35m"
+        colorcodes['bcyan']      = monochrome_logs ? '' : "\033[1;36m"
+        colorcodes['bwhite']     = monochrome_logs ? '' : "\033[1;37m"
+
+        // Underline
+        colorcodes['ublack']     = monochrome_logs ? '' : "\033[4;30m"
+        colorcodes['ured']       = monochrome_logs ? '' : "\033[4;31m"
+        colorcodes['ugreen']     = monochrome_logs ? '' : "\033[4;32m"
+        colorcodes['uyellow']    = monochrome_logs ? '' : "\033[4;33m"
+        colorcodes['ublue']      = monochrome_logs ? '' : "\033[4;34m"
+        colorcodes['upurple']    = monochrome_logs ? '' : "\033[4;35m"
+        colorcodes['ucyan']      = monochrome_logs ? '' : "\033[4;36m"
+        colorcodes['uwhite']     = monochrome_logs ? '' : "\033[4;37m"
+
+        // High Intensity
+        colorcodes['iblack']     = monochrome_logs ? '' : "\033[0;90m"
+        colorcodes['ired']       = monochrome_logs ? '' : "\033[0;91m"
+        colorcodes['igreen']     = monochrome_logs ? '' : "\033[0;92m"
+        colorcodes['iyellow']    = monochrome_logs ? '' : "\033[0;93m"
+        colorcodes['iblue']      = monochrome_logs ? '' : "\033[0;94m"
+        colorcodes['ipurple']    = monochrome_logs ? '' : "\033[0;95m"
+        colorcodes['icyan']      = monochrome_logs ? '' : "\033[0;96m"
+        colorcodes['iwhite']     = monochrome_logs ? '' : "\033[0;97m"
+
+        // Bold High Intensity
+        colorcodes['biblack']    = monochrome_logs ? '' : "\033[1;90m"
+        colorcodes['bired']      = monochrome_logs ? '' : "\033[1;91m"
+        colorcodes['bigreen']    = monochrome_logs ? '' : "\033[1;92m"
+        colorcodes['biyellow']   = monochrome_logs ? '' : "\033[1;93m"
+        colorcodes['biblue']     = monochrome_logs ? '' : "\033[1;94m"
+        colorcodes['bipurple']   = monochrome_logs ? '' : "\033[1;95m"
+        colorcodes['bicyan']     = monochrome_logs ? '' : "\033[1;96m"
+        colorcodes['biwhite']    = monochrome_logs ? '' : "\033[1;97m"
+
+        return colorcodes
+    }
+
+    //
+    // Remove all ANSII colors from a piece of text
+    //
+    public static String removeColors(String input) {
+        if (!input) {return input}
+        String output = input
+        List colors = logColors(false).collect { it.value }
+        colors.each { color ->
+            output = output.replace(color, "")
+        }
+        return output
+    }
+}
