+++ plugins/nf-schema/src/main/nextflow/validation/utils/Common.groovy
@@ -1,0 +1,67 @@
+package nextflow.validation.utils
+
+import org.yaml.snakeyaml.Yaml
+import org.json.JSONArray
+import org.json.JSONObject
+import org.json.JSONPointer
+import org.json.JSONPointerException
+import nextflow.Global
+
+import groovy.json.JsonGenerator
+import groovy.json.JsonSlurper
+import groovy.util.logging.Slf4j
+import java.nio.file.Path
+
+import nextflow.validation.exceptions.SchemaValidationException
+
+/**
+ * A collection of commonly used functions
+ *
+ * @author : mirpedrol <mirp.julia@gmail.com>
+ * @author : nvnieuwk <nicolas.vannieuwkerke@ugent.be>
+ * @author : KevinMenden
+ */
+
+@Slf4j
+public class Common {
+
+    //
+    // Resolve Schema path relative to main workflow directory
+    //
+    public static String getSchemaPath(String baseDir, String schemaFilename) {
+        if (Path.of(schemaFilename).exists()) {
+            return schemaFilename
+        } else {
+            return "${baseDir}/${schemaFilename}"
+        }
+    }
+
+    //
+    // Function to get the value from a JSON pointer
+    //
+    public static Object getValueFromJson(String jsonPointer, Object json) {
+        def JSONPointer schemaPointer = new JSONPointer(jsonPointer)
+        try {
+            return schemaPointer.queryFrom(json) ?: ""
+        } catch (JSONPointerException e) {
+            return ""
+        }
+    }
+
+    //
+    // Get maximum number of characters across all parameter names
+    //
+    public static Integer paramsMaxChars( Map paramsMap ) {
+        return Collections.max(paramsMap.collect { _, val -> 
+            def Map groupParams = val as Map
+            longestStringLength(groupParams.keySet() as List<String> )
+        })
+    }
+
+    //
+    // Get the size of the longest string value in a list of strings
+    //
+    public static Integer longestStringLength( List<String> strings ) {
+        return strings ? Collections.max(strings.collect { it.size() }) : 0
+    }
+}
