+++ plugins/nf-schema/src/main/nextflow/validation/utils/Files.groovy
@@ -1,4 +1,4 @@
-package nextflow.validation
+package nextflow.validation.utils
 
 import org.yaml.snakeyaml.Yaml
 import org.json.JSONArray
@@ -13,17 +13,23 @@ import groovy.util.logging.Slf4j
 import java.nio.file.Path
 
 import nextflow.validation.exceptions.SchemaValidationException
+import static nextflow.validation.utils.Common.getValueFromJson
+import static nextflow.validation.utils.Types.castToType
 
 /**
+ * A collection of functions used to get data from files
+ *
  * @author : mirpedrol <mirp.julia@gmail.com>
  * @author : nvnieuwk <nicolas.vannieuwkerke@ugent.be>
  * @author : KevinMenden
  */
 
 @Slf4j
-public class Utils {
+public class Files {
 
-    // Function to infer the file type of a samplesheet
+    //
+    // Function to detect if a file is a CSV, TSV, JSON or YAML file
+    //
     public static String getFileType(Path file) {
         def String extension = file.getExtension()
         if (extension in ["csv", "tsv", "yml", "yaml", "json"]) {
@@ -46,16 +52,20 @@ public class Utils {
         }
     }
 
+    //
     // Function to get the header from a CSV or TSV file
+    //
     public static String getHeader(Path file) {
         def String header
         file.withReader { header = it.readLine() }
         return header
     }
 
-    // Converts a given file to a List
+    //
+    // Converts a given file to an Groovy object (either a List or a Map)
+    //
     public static Object fileToObject(Path file, Path schema) {
-        def String fileType = Utils.getFileType(file)
+        def String fileType = getFileType(file)
         def String delimiter = fileType == "csv" ? "," : fileType == "tsv" ? "\t" : null
         def Map schemaMap = (Map) new JsonSlurper().parse( schema )
         def Map types = variableTypes(schema)
@@ -89,10 +99,11 @@ public class Utils {
         }
     }
 
-    // Converts a given file to a JSONArray
+    //
+    // Converts a given file to a JSON type (either JSONArray or JSONObject)
+    //
     public static Object fileToJson(Path file, Path schema) {
         // Remove all null values from JSON object
-        // and convert the groovy object to a JSONArray
         def jsonGenerator = new JsonGenerator.Options()
             .excludeNulls()
             .build()
@@ -108,46 +119,8 @@ public class Utils {
     }
 
     //
-    // Cast a value to the provided type in a Strict mode
+    // Get a map that contains the type for each key in a JSON schema file
     //
-
-    public static Object castToType(Object input) {
-        def Set<String> validBooleanValues = ['true', 'false'] as Set
-
-        if (input instanceof Map) {
-            // Cast all values in the map
-            def Map output = [:]
-            input.each { k, v ->
-                output[k] = castToType(v)
-            }
-            return output
-        }
-        else if (input instanceof List) {
-            // Cast all values in the list
-            def List output = []
-            for( entry : input ) {
-                output.add(castToType(entry))
-            }
-            return output
-        } else if (input instanceof String) {
-            // Cast the string if there is one
-            if (input == "") {
-                return null
-            }
-            return JSONObject.stringToValue(input)
-        }
-    }
-
-    // Resolve Schema path relative to main workflow directory
-    public static String getSchemaPath(String baseDir, String schemaFilename) {
-        if (Path.of(schemaFilename).exists()) {
-            return schemaFilename
-        } else {
-            return "${baseDir}/${schemaFilename}"
-        }
-    }
-
-    // Function to obtain the variable types of properties from a JSON Schema
     public static Map variableTypes(Path schema) {
         def Map variableTypes = [:]
         def String type = ''
@@ -176,124 +149,6 @@ public class Utils {
         }
 
         return variableTypes
-    }
-
-    // Function to check if a String value is an Integer
-    public static Boolean isInteger(String input) {
-        try {
-            input as Integer
-            return true
-        } catch (NumberFormatException e) {
-            return false
-        }
-    }
-
-    // Function to check if a String value is a Float
-    public static Boolean isFloat(String input) {
-        try {
-            input as Float
-            return true
-        } catch (NumberFormatException e) {
-            return false
-        }
-    }
-
-    // Function to check if a String value is a Double
-    public static Boolean isDouble(String input) {
-        try {
-            input as Double
-            return true
-        } catch (NumberFormatException e) {
-            return false
-        }
-    }
-
-    // Function to get the value from a JSON pointer
-    public static Object getValueFromJson(String jsonPointer, Object json) {
-        def JSONPointer schemaPointer = new JSONPointer(jsonPointer)
-        try {
-            return schemaPointer.queryFrom(json) ?: ""
-        } catch (JSONPointerException e) {
-            return ""
-        }
-    }
-
-    //
-    // ANSII Colours used for terminal logging
-    //
-    public static Map logColours(Boolean monochrome_logs) {
-        Map colorcodes = [:]
-
-        // Reset / Meta
-        colorcodes['reset']      = monochrome_logs ? '' : "\033[0m"
-        colorcodes['bold']       = monochrome_logs ? '' : "\033[1m"
-        colorcodes['dim']        = monochrome_logs ? '' : "\033[2m"
-        colorcodes['underlined'] = monochrome_logs ? '' : "\033[4m"
-        colorcodes['blink']      = monochrome_logs ? '' : "\033[5m"
-        colorcodes['reverse']    = monochrome_logs ? '' : "\033[7m"
-        colorcodes['hidden']     = monochrome_logs ? '' : "\033[8m"
-
-        // Regular Colors
-        colorcodes['black']      = monochrome_logs ? '' : "\033[0;30m"
-        colorcodes['red']        = monochrome_logs ? '' : "\033[0;31m"
-        colorcodes['green']      = monochrome_logs ? '' : "\033[0;32m"
-        colorcodes['yellow']     = monochrome_logs ? '' : "\033[0;33m"
-        colorcodes['blue']       = monochrome_logs ? '' : "\033[0;34m"
-        colorcodes['purple']     = monochrome_logs ? '' : "\033[0;35m"
-        colorcodes['cyan']       = monochrome_logs ? '' : "\033[0;36m"
-        colorcodes['white']      = monochrome_logs ? '' : "\033[0;37m"
-
-        // Bold
-        colorcodes['bblack']     = monochrome_logs ? '' : "\033[1;30m"
-        colorcodes['bred']       = monochrome_logs ? '' : "\033[1;31m"
-        colorcodes['bgreen']     = monochrome_logs ? '' : "\033[1;32m"
-        colorcodes['byellow']    = monochrome_logs ? '' : "\033[1;33m"
-        colorcodes['bblue']      = monochrome_logs ? '' : "\033[1;34m"
-        colorcodes['bpurple']    = monochrome_logs ? '' : "\033[1;35m"
-        colorcodes['bcyan']      = monochrome_logs ? '' : "\033[1;36m"
-        colorcodes['bwhite']     = monochrome_logs ? '' : "\033[1;37m"
-
-        // Underline
-        colorcodes['ublack']     = monochrome_logs ? '' : "\033[4;30m"
-        colorcodes['ured']       = monochrome_logs ? '' : "\033[4;31m"
-        colorcodes['ugreen']     = monochrome_logs ? '' : "\033[4;32m"
-        colorcodes['uyellow']    = monochrome_logs ? '' : "\033[4;33m"
-        colorcodes['ublue']      = monochrome_logs ? '' : "\033[4;34m"
-        colorcodes['upurple']    = monochrome_logs ? '' : "\033[4;35m"
-        colorcodes['ucyan']      = monochrome_logs ? '' : "\033[4;36m"
-        colorcodes['uwhite']     = monochrome_logs ? '' : "\033[4;37m"
-
-        // High Intensity
-        colorcodes['iblack']     = monochrome_logs ? '' : "\033[0;90m"
-        colorcodes['ired']       = monochrome_logs ? '' : "\033[0;91m"
-        colorcodes['igreen']     = monochrome_logs ? '' : "\033[0;92m"
-        colorcodes['iyellow']    = monochrome_logs ? '' : "\033[0;93m"
-        colorcodes['iblue']      = monochrome_logs ? '' : "\033[0;94m"
-        colorcodes['ipurple']    = monochrome_logs ? '' : "\033[0;95m"
-        colorcodes['icyan']      = monochrome_logs ? '' : "\033[0;96m"
-        colorcodes['iwhite']     = monochrome_logs ? '' : "\033[0;97m"
-
-        // Bold High Intensity
-        colorcodes['biblack']    = monochrome_logs ? '' : "\033[1;90m"
-        colorcodes['bired']      = monochrome_logs ? '' : "\033[1;91m"
-        colorcodes['bigreen']    = monochrome_logs ? '' : "\033[1;92m"
-        colorcodes['biyellow']   = monochrome_logs ? '' : "\033[1;93m"
-        colorcodes['biblue']     = monochrome_logs ? '' : "\033[1;94m"
-        colorcodes['bipurple']   = monochrome_logs ? '' : "\033[1;95m"
-        colorcodes['bicyan']     = monochrome_logs ? '' : "\033[1;96m"
-        colorcodes['biwhite']    = monochrome_logs ? '' : "\033[1;97m"
-
-        return colorcodes
-    }
-
-    public static String removeColors(String input) {
-        if (!input) {return input}
-        String output = input
-        List colors = logColours(false).collect { it.value }
-        colors.each { color ->
-            output = output.replace(color, "")
-        }
-        return output
     }
 
     //
@@ -372,22 +227,5 @@ public class Utils {
         }
 
         return paramsMap
-    }
-
-    //
-    // Get maximum number of characters across all parameter names
-    //
-    public static Integer paramsMaxChars( Map paramsMap ) {
-        return Collections.max(paramsMap.collect { _, val -> 
-            def Map groupParams = val as Map
-            longestStringLength(groupParams.keySet() as List<String> )
-        })
-    }
-
-    //
-    // Get the size of the longest string value in a list of strings
-    //
-    public static Integer longestStringLength( List<String> strings ) {
-        return strings ? Collections.max(strings.collect { it.size() }) : 0
     }
 }
